// fringe/types/api/locationDto.ts
export interface LocationDto {
    locationId: number;
    locationName: string;
    name: string;
    id: string;
    address: string;
    suburb: string;
    state: string;
    postalCode: string;
    country: string;
    latitude: number;
    longitude: number;
    parkingAvailable: boolean;
    active: boolean;
}